massive doxygenation of audiere.h changed volumes to be normalized floats made OpenSound take a sample source instead of a filename/file
git-svn-id: http://audiere.svn.sourceforge.net/svnroot/audiere/trunk@233 cb5fa4c3-b70d-0410-b4f6-90f5898e5689

+++ src/device_ds.h
@@ -70,8 +70,9 @@ namespace audiere {
     void stop();
     void reset();
     bool isPlaying();
-    void setVolume(int volume);
-    int  getVolume();
+
+    void  setVolume(float volume);
+    float getVolume();
 
     void fillStream();
     void update();
@@ -91,7 +92,7 @@ namespace audiere {
     SampleSource* m_source;
     int m_sample_size;  // convenience: bits per sample * channel count / 8
 
-    int m_volume;
+    float m_volume;
     ::BYTE* m_last_sample; // the last sample read (used for clickless silence)
 
     friend class DSAudioDevice;
